pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "hybrid-api-scraping-collector-template"
version = "0.1.0"
description = "Template for combining API data collection and HTML scraping into a unified dataset."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
dependencies = [
    "requests",
    "beautifulsoup4",
    "PyYAML",
    "python-dotenv",
    "pandas",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.pytest.ini_options]
pythonpath = ["src"]